test: Move tests into the correct file
Refs: BLAIS5-3118
